Decode subprocess output in UTF-8...
and raise Python requirement to 3.6.

A Windows user with cp936 (i.e. GBK) as locale.getpreferredencoding() reported
error when ffmpeg output is being decoded. Only Python 3.6 allows us to pass
encoding and errors to subprocess.Popen in universal newlines mode,[1] though,
so we have to raise Python requirement to 3.6 for this to work.

[1] https://docs.python.org/3/library/subprocess.html#subprocess.Popen