Fix default downloads dir handling
It was completely, utterly broken, so it's hard to say what changed,
because everything changed. Suffice to say it's much more robust now.

Also, since the worker process is forked before we even load the
configs, using the global Config in the worker process doesn't
work. Now we're passing the downloads dir as outputFile to the worker
process when no output file is specified.

To prevent further misuse of Config, we now use an instance in the
'config' attribute MainWindow instead of relying on class variables of
Config, which have now been removed.