Restore monospaced font in messages box
Monospaced font was silently lost when we switched to rich text. Setting
the font-family property in CSS seems to be the only way to restore it,
and unfortunately, 'monospace' isn't a supported font, so I have to hard
code Courier.

Despite my efforts, the font is just too large on Windows 7 (haven't
tested on Windows 10 yet), and I can't figure out a way to reduce the
font size -- font-size in CSS doesn't change a thing (it works under
macOS); neither does QTextEdit::setFontPointSize.